+++ docsfs.go
@@ -1,0 +1,6 @@
+package docsfs
+
+import "embed"
+
+//go:embed docs
+var FS embed.FS